feat: Add SQL script to create states and cities tables
This commit adds a new SQL script `14-model_city_fetch_by_state.sql` that creates the `states` and `cities` tables in the `hbtn_0e_14_usa` database. It also inserts some initial data into the tables. This script is necessary for the object-relational mapping functionality implemented in recent user commits.

The commit message follows the established convention of describing the purpose of the script being added.